Fix write_reg_int test assertions
Corrected the expected byte arrays in `test_write_reg_int` to match the actual behavior of `write_reg_int` and `winreg`.
- `REG_NONE` and other 32-bit types now expect `u32` byte arrays (4 bytes).
- `REG_QWORD` now expects a `u64` byte array (8 bytes).
This resolves test failures where 4-byte registry values were being compared against 8-byte expectations and vice-versa.

+++ implants/lib/eldritchv2/stdlib/eldritch-libsys/src/std/write_reg_int_impl.rs
@@ -152,7 +152,7 @@ mod tests {
             hkcu = RegKey::predef(HKEY_CURRENT_USER);
             subky = hkcu.open_subkey(format!("SOFTWARE\\{}", id))?;
             val2 = subky.get_raw_value("FOO2")?;
-            assert_eq!(val2.bytes, 12345678u64.to_le_bytes().to_vec());
+            assert_eq!(val2.bytes, 12345678u32.to_le_bytes().to_vec());
 
             //delete temp regkey
             hkcu.delete_subkey(format!("SOFTWARE\\{}", id))?;
@@ -312,7 +312,7 @@ mod tests {
             hkcu = RegKey::predef(HKEY_CURRENT_USER);
             subky = hkcu.open_subkey(format!("SOFTWARE\\{}", id))?;
             val2 = subky.get_raw_value("FOO2")?;
-            assert_eq!(val2.bytes, 12345678u32.to_le_bytes().to_vec());
+            assert_eq!(val2.bytes, 12345678u64.to_le_bytes().to_vec());
 
             //delete temp regkey
             hkcu.delete_subkey(format!("SOFTWARE\\{}", id))?;
